ref: support multiple worktrees check for refs
We have already set up the infrastructure to check the consistency for
refs, but we do not support multiple worktrees. As we decide to add more
checks for ref content, we need to set up support for multiple
worktrees.

We need to display the full check target name such as
"worktrees/<id>/refs/worktree/foo", so we should know the id of the
worktree to get the full target name. Add a new parameter "struct
worktree *" for "refs-internal.h::fsck_fn". Then, change the related
interfaces to follow this new interface.

Because we should only check once for "packed-refs", use
"is_main_worktree" method to skip checking "pack-refs" when the
worktree is not the main worktree.

Then, enhance the "files-backend.c::files_fsck_refs_dir" function to add
"worktree/<id>/" prefix when we are not in the main worktree.

Last, add new test to check the refname when there are multiple
worktrees to exercise the code.

Mentored-by: Patrick Steinhardt <ps@pks.im>
Mentored-by: Karthik Nayak <karthik.188@gmail.com>
Signed-off-by: shejialuo <shejialuo@gmail.com>