ref-cache: replace "FLEX_ALLOC_MEM" with "FELX_ALLOC_STR"
We use "FLEX_ALLOC_MEM" macro to initialize the "name" string field of
"struct ref_enrty" in "create_dir_entry" function. This macro needs the
callers to explicitly specify the memory size. So, we let the callers
pass the "len" to "create_dir_entry" function. However, when the callers
use the raw string for "dirname", it's bad that the callers need to
calculate the length of the raw string and pass the literal to the
"create_dir_entry" function like the following shows:

  add_entry_to_dir(dir, create_dir_entry(refs->loose, "refs/", 5));

In "create_ref_entry", we use "FELX_ALLOC_STR" macro to create a "struct
ref_entry *" for a loose ref. And we also return the "struct ref_enrty *"
in "create_dir_entry" function, we'd better also align with this
behavior.

Replace "FLEX_ALLOC_MEM" with "FELX_ALLOC_STR" in "create_dir_entry".
Then, remove the "len" parameter in "create_dir_entry" function.

Signed-off-by: shejialuo <shejialuo@gmail.com>